[MOVE] : AuthRepository / Auth Model 도메인 모듈로 이동 #415
