+++ app/src/main/java/org/sopt/official/data/repository/AuthRepositoryImpl.kt
@@ -26,12 +26,12 @@ package org.sopt.official.data.repository
 
 import org.sopt.official.data.mapper.AuthMapper
 import org.sopt.official.data.model.request.LogOutRequest
-import org.sopt.official.data.model.response.LogOutResponse
 import org.sopt.official.data.source.api.auth.LocalAuthDataSource
 import org.sopt.official.data.source.api.auth.RemoteAuthDataSource
-import org.sopt.official.domain.entity.auth.Token
-import org.sopt.official.domain.entity.auth.UserStatus
-import org.sopt.official.domain.repository.AuthRepository
+import domain.model.Token
+import domain.model.UserStatus
+import domain.repository.AuthRepository
+import domain.model.LogOut
 import org.sopt.official.network.model.request.RefreshRequest
 import javax.inject.Inject
 
@@ -57,13 +57,13 @@ class AuthRepositoryImpl @Inject constructor(
 
     override suspend fun logout(
         pushToken: String
-    ): Result<LogOutResponse> = runCatching {
+    ): Result<LogOut> = runCatching {
         remoteAuthDataSource.logout(
             LogOutRequest(
                 platform = "Android",
                 pushToken = pushToken
             )
-        )
+        ).toEntity()
     }
 
     override suspend fun clearLocalData() {
