+++ app/src/main/java/org/sopt/official/data/source/impl/DefaultLocalAuthDataSource.kt
@@ -26,8 +26,8 @@ package org.sopt.official.data.source.impl
 
 import org.sopt.official.network.persistence.SoptDataStore
 import org.sopt.official.data.source.api.auth.LocalAuthDataSource
-import org.sopt.official.domain.entity.auth.Token
-import org.sopt.official.domain.entity.auth.UserStatus
+import domain.model.Token
+import domain.model.UserStatus
 import javax.inject.Inject
 
 class DefaultLocalAuthDataSource @Inject constructor(
