+++ app/src/main/java/org/sopt/official/feature/auth/AuthActivity.kt
@@ -53,7 +53,7 @@ import org.sopt.official.config.messaging.SoptFirebaseMessagingService.Companion
 import org.sopt.official.data.model.request.AuthRequest
 import org.sopt.official.data.service.AuthService
 import org.sopt.official.databinding.ActivityAuthBinding
-import org.sopt.official.domain.entity.auth.UserStatus
+import domain.model.UserStatus
 import org.sopt.official.feature.home.HomeActivity
 import org.sopt.official.network.persistence.SoptDataStore
 import org.sopt.official.util.dp
