+++ app/src/main/java/org/sopt/official/util/Auth.kt
@@ -25,9 +25,9 @@
 package org.sopt.official.util
 
 import org.sopt.official.network.model.response.OAuthToken
-import org.sopt.official.domain.entity.auth.Auth
-import org.sopt.official.domain.entity.auth.Token
-import org.sopt.official.domain.entity.auth.UserStatus
+import domain.model.Auth
+import domain.model.Token
+import domain.model.UserStatus
 
 fun OAuthToken.toEntity() = Auth(
     Token(
