+++ build.gradle.kts
@@ -33,6 +33,7 @@ plugins {
     alias(libs.plugins.junit5) apply false
     alias(libs.plugins.app.distribution) apply false
     alias(libs.plugins.spotless) apply false
+    alias(libs.plugins.org.jetbrains.kotlin.jvm) apply false
 }
 
 subprojects {
