+++ domain/mypage/src/main/kotlin/domain/model/Auth.kt
@@ -22,17 +22,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.domain.entity.auth
+package domain.model
 
-enum class UserStatus(
-    val value: String
-) {
-    ACTIVE("ACTIVE"),
-    INACTIVE("INACTIVE"),
-    UNAUTHENTICATED("UNAUTHENTICATED");
-
-    companion object {
-        fun of(value: String) = entries.find { it.value == value }
-            ?: throw IllegalArgumentException("Invalid user status: $value")
-    }
-}
+data class Auth(
+    val token: Token,
+    val status: UserStatus
+)
