+++ domain/mypage/src/main/kotlin/domain/model/LogOut.kt
@@ -1,0 +1,7 @@
+package domain.model
+
+
+data class LogOut(
+    val success: Boolean,
+    val message: String
+)
