+++ domain/mypage/src/main/kotlin/domain/model/Token.kt
@@ -22,9 +22,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.domain.entity.auth
+package domain.model
 
-data class Auth(
-    val token: Token,
-    val status: UserStatus
+data class Token(
+    val accessToken: String,
+    val refreshToken: String,
+    val playgroundToken: String
 )
